Implement MyoObjectFactory
The MyoObjectFactory now has code inside of it that exists and does
things. Namely, it covers all the static calls to creation methods in
MyoSharp.

The MyoDal can rely on this to set up the connection objects now,
meaning it is testable. Meaning it now passes the tests I just wrote for
it!